Split out getCurrentPlan
